+++ datatransfer/network/libp2p_impl.go
@@ -6,6 +6,7 @@ import (
 	"io"
 	"time"
 
+	ggio "github.com/gogo/protobuf/io"
 	logging "github.com/ipfs/go-log"
 	"github.com/libp2p/go-libp2p-core/helpers"
 	"github.com/libp2p/go-libp2p-core/host"
